feat: create cache provider
