+++ pages/_app.tsx
@@ -1,0 +1,4 @@
+import { App } from '@application';
+
+// eslint-disable-next-line unicorn/prefer-export-from
+export default App;
