+++ src/application/app/app.component.tsx
@@ -1,0 +1,9 @@
+import { AppProvider } from '@application';
+
+import { IAppProps } from './app.types';
+
+export const App = ({ Component, pageProps }: IAppProps) => (
+  <AppProvider>
+    <Component {...pageProps} />
+  </AppProvider>
+);
