+++ src/application/app/app.types.ts
@@ -1,0 +1,6 @@
+import { EmotionCache } from '@emotion/react';
+import { AppProps } from 'next/app';
+
+export interface IAppProps extends AppProps {
+  emotionCache?: EmotionCache;
+}
