+++ src/application/index.ts
@@ -1,1 +1,3 @@
 export * from './providers';
+
+export * from './app/app.component';
