+++ src/application/providers/app-provider/app-provider.component.tsx
@@ -1,8 +1,10 @@
-import { LocaleProvider, ThemeProvider } from '..';
+import { CacheProvider, LocaleProvider, ThemeProvider } from '..';
 import { IAppProviderProps } from './app-provider.types';
 
 export const AppProvider = ({ children }: IAppProviderProps): JSX.Element => (
   <ThemeProvider>
-    <LocaleProvider>{children}</LocaleProvider>
+    <LocaleProvider>
+      <CacheProvider>{children}</CacheProvider>
+    </LocaleProvider>
   </ThemeProvider>
 );
