+++ src/application/providers/cache-provider/cache-provider.component.tsx
@@ -1,0 +1,11 @@
+import { CacheProvider as EmotionCacheProvider } from '@emotion/react';
+
+import { createEmotionCache } from '@utils';
+
+import { ICacheProviderProps } from './cache-provider.types';
+
+const clientSideEmotionCache = createEmotionCache();
+
+export const CacheProvider = ({ children }: ICacheProviderProps): JSX.Element => (
+  <EmotionCacheProvider value={clientSideEmotionCache}>{children}</EmotionCacheProvider>
+);
