+++ src/application/providers/cache-provider/cache-provider.types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface ICacheProviderProps {
+  children: ReactNode | ReactNode[];
+}
