+++ src/application/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './app-provider/app-provider.component';
+export * from './cache-provider/cache-provider.component';
 export * from './theme-provider/theme-provider.component';
 export * from './locale-provider/locale-provider.component';
